Added config.json that stores server IPAddress and port. Every part of project is containerized (nginx frontend, golang backend, postgresql, pgadmin(added now) ). The whole project now in docker-compose and can be build using "docker-compose up --build" command (also creates db and tables inside it using init.sql). Added pgadmin that is accessed via browser. Added .env file that stores environmental variables that are used to determine ports, passwords, addresses, etc. Both frontend nginx and backend golang now use TLS to communicate via https. Changed wastebing image to more lightweight one.
