+++ .env
@@ -1,0 +1,23 @@
+# Postgres Live
+DB_HOST=todobukh-postgres
+# DB_HOST=127.0.0.1                 # when running the app without docker 
+DB_DRIVER=postgres
+API_SECRET=bukh           # Used for creating a JWT. Can be anything 
+DB_USER=postgres
+DB_PASSWORD=postgres
+DB_NAME=todobukh
+DB_PORT=5432
+
+# Used by pgadmin service 
+PGADMIN_DEFAULT_EMAIL=live@admin.com
+PGADMIN_DEFAULT_PASSWORD=password
+
+# # Postgres Test
+# TEST_DB_HOST=postgres_test                      
+# # TEST_DB_HOST=127.0.0.1                        # when running the app without docker 
+# TEST_DB_DRIVER=postgres
+# TEST_API_SECRET=98hbun98h
+# TEST_DB_USER=steven
+# TEST_DB_PASSWORD=password
+# TEST_DB_NAME=fullstack_api_test
+# TEST_DB_PORT=5432
