+++ go.mod
@@ -1,11 +1,10 @@
-module main.go
+module todobukh
 
-go 1.21.1
+go 1.21.5
 
 require (
-	github.com/go-chi/chi/v5 v5.0.10
-	github.com/go-chi/cors v1.2.1
-	github.com/golang-jwt/jwt/v5 v5.0.0
+	github.com/go-chi/chi/v5 v5.0.11
+	github.com/golang-jwt/jwt/v5 v5.2.0
 	github.com/lib/pq v1.10.9
-	golang.org/x/crypto v0.13.0
+	golang.org/x/crypto v0.18.0
 )
