+++ ssl/localhost.crt
@@ -1,0 +1,21 @@
+-----BEGIN CERTIFICATE-----
+MIIDazCCAlOgAwIBAgIUBX+ln5HTpYIcznsCmq6ppiUSZEwwDQYJKoZIhvcNAQEL
+BQAwRTELMAkGA1UEBhMCQVUxEzARBgNVBAgMClNvbWUtU3RhdGUxITAfBgNVBAoM
+GEludGVybmV0IFdpZGdpdHMgUHR5IEx0ZDAeFw0yNDAxMjEwOTQzMTBaFw0yNTAx
+MjAwOTQzMTBaMEUxCzAJBgNVBAYTAkFVMRMwEQYDVQQIDApTb21lLVN0YXRlMSEw
+HwYDVQQKDBhJbnRlcm5ldCBXaWRnaXRzIFB0eSBMdGQwggEiMA0GCSqGSIb3DQEB
+AQUAA4IBDwAwggEKAoIBAQDUu0GNiidLo5nY4tBvVvhRnD602tJdbBYTr1XdA6Wg
+oRYmqeknMrV+Ksafmz/dUbVeAb8s2Auw60CBrbhAtCJH16WEBhOwW0BhdV3bakq1
++wxp6Dg0/MNHuczglXHgVGwcXGSTd+Yn2ORKzlzY3wEyOFHs/7yB0q9VRfjac1c/
+VhI8NeHt76drTG2BgSs36bZgnPPMUoic0rYo/yAwMd9eIpFckx8KRUediJULI8uI
+POwizHuM0+nVcbT3WS1GTK+LHxDXjUn06odXKXiUrC1VcDwg/EflQViWH0wmO3LS
+hKM2PVztCJwK3efN9ENPjID5qQomGP+ETVTG1q6FgrgLAgMBAAGjUzBRMB0GA1Ud
+DgQWBBSPyXNiSgzkdshT0ZOsrKLReWmsgTAfBgNVHSMEGDAWgBSPyXNiSgzkdshT
+0ZOsrKLReWmsgTAPBgNVHRMBAf8EBTADAQH/MA0GCSqGSIb3DQEBCwUAA4IBAQDE
+x3gygJkH/MiyWb5fq7aubC43z2qm7unYz786pyA/He/b8t98hdMA5D1ff9q+desf
+ZxwngdkY1zVA9JIImJKaxJ2vw/Cl6Tcd3R6X7m08WHCcDSmkRlaETcBKhfmVx2n1
+I8ELpaTR3xxWZHWKXjlwZIrHRwK2ao5DLA1PDnj0i6+psSVFCi4fwN80acA/LGJK
+ZMFlF/XKnxbuOPhKFptkZorD69GuLU+ENPGn0DtTSPlOQ1Bot0DKbBb81GRtkc+0
+7hFO/F+QSSMP0clw7A6oIUkl0GvkaXqq2zg9WHHSzaFzYHBO/1AshjCDax2Kp845
+7oZ78QxGs5kCzyJWiSRa
+-----END CERTIFICATE-----
