+++ ssl/localhost.key
@@ -1,0 +1,28 @@
+-----BEGIN PRIVATE KEY-----
+MIIEvAIBADANBgkqhkiG9w0BAQEFAASCBKYwggSiAgEAAoIBAQDUu0GNiidLo5nY
+4tBvVvhRnD602tJdbBYTr1XdA6WgoRYmqeknMrV+Ksafmz/dUbVeAb8s2Auw60CB
+rbhAtCJH16WEBhOwW0BhdV3bakq1+wxp6Dg0/MNHuczglXHgVGwcXGSTd+Yn2ORK
+zlzY3wEyOFHs/7yB0q9VRfjac1c/VhI8NeHt76drTG2BgSs36bZgnPPMUoic0rYo
+/yAwMd9eIpFckx8KRUediJULI8uIPOwizHuM0+nVcbT3WS1GTK+LHxDXjUn06odX
+KXiUrC1VcDwg/EflQViWH0wmO3LShKM2PVztCJwK3efN9ENPjID5qQomGP+ETVTG
+1q6FgrgLAgMBAAECggEAG6P7uoD5kMcVwFaIL/n5kzEH/MZ0kEBGW2cWyFF8Ux4F
+xh729RQ4/McLaRrqS2lzIAnTZeeUbU0ZOUBHMTEL8JBr0yGvATp6ElpbnZzL+aVA
+Uap4qHncgxSjMt3NdBEZaFNLm4BNXrQRxZ6BSzR+mqCIFJE4qcL5lneHBBX51M1a
+GwdZCYJG22qckfYR2lK69Mi3HR1lYjuSiJ84BscEdsRUVUEV9O7pl3upsIZGYsvz
+ahY4CwuPMvqi2MrLKQyFkIfqHUHAEbZmYFxaPs7R3WIdYW87U2zcxEyKvDpdYqWa
+HNljIC2knvE7uq4m4yy4PADwPq7OgduRCRvXDyymoQKBgQD3cmUAbXH22LcuzTKQ
+XiJAGRyK3kVK4R8uZ24XWhwLKCFQuI4CTAQPZYg7r9cXE/i8q2HF7lLdqrR0Rg++
+kp4EbTqZ/LK11ND91Xn3wONFqgvNshEDaV1kkTofuF6MF/QML7k9ORk65KzpYF0B
+5qCJsiSBgSYZ6DKuTMXApP2+mQKBgQDcFawfEOqMYq2JXVgrrzY8wKWhp5Z5xV99
+ctPvqczwHcUc1nGa1B1jZeFommnwctVOIUlBRpxRVjIk/cJJ7dWh2HtNl7x+0Qn2
+LEiWhKp7CMUXpMtWXKiusrPbxqMcaWJKHBObdPqt6d5W4JOzc1sElfKtu+K6I14B
+bax45lxGQwKBgEUKyB+8aeYERehrHManXUshNpOQvSjNc5PRDCi73cNHKTSKqiT9
+7MwW+lIdGBi6W9bYwuklPD3ZWYxv2AjGhAycJZO9rGeLv9xzahq02TUg4zY7jv+y
+sHpVAzDCkq7oKBIWZsssWVvTXmOnnYvgO77v07vGu0jTPwDfLvY3kw1RAoGAauhQ
+KusRtE3+1Do2qUm8Q/ikvIjm7JH2175euw21knIznNWCa95bxDvqKeDQU0/E0+YR
+apb9Oflv9E4cAcN5Qdwh0UhC9Stf9mf+frdFBOz2qIyRhYsPGYWPHtQ3yYZbi3pl
+/97sImz3ZyZCOSpw7SJVtxUMhrTFZlY7MKNd8XcCgYByyaInBAVd9rdBnGfxnCL8
+nuCNVdehAVxEYnDvi/s0jFIGcCpjlKMZ5OjF+V+C0PiITESZJquMI2yzwlODNLud
+koQxC3PfC+Za4vPz3PGyUoPtoMqe0oRP/7zaOzCdNhlBzaz+iSp+Wyw2QT24FsGM
+nuld34RtxZM2iDz4qo4xlA==
+-----END PRIVATE KEY-----
