+++ www/todobukh.ru/config.json
@@ -1,0 +1,5 @@
+{
+    "ipAddress": "localhost",
+    "port": 3000
+}
+
